# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supermarket API",
        "description": "API for managing a supermarket with departments, products, and sales",
        "contact": {
            "name": "Support Team",
            "email": "admin@supermarket.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products",
                "description": "Search products by name, category, or rayon_id",
                "operationId": "33867295f3d0876f2cb47a1e6da11e02",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Search by product name (partial match)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by product category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rayon_id",
                        "in": "query",
                        "description": "Filter by rayon ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "rayon_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Milk"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Dairy"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 2.99
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "stock_threshold": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "is_popular": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "is_on_sale": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "sale_price": {
                                                "type": "number",
                                                "example": 1.99,
                                                "nullable": true
                                            },
                                            "rayon": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dairy Section"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/popular-or-on-sale": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get popular or on-sale products",
                "description": "Retrieve products that are popular, on sale, or both",
                "operationId": "7604dd1f186f955847c25f0e6d3412f2",
                "parameters": [
                    {
                        "name": "popular",
                        "in": "query",
                        "description": "Filter by popular products (true/false)",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "on_sale",
                        "in": "query",
                        "description": "Filter by on-sale products (true/false)",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "rayon_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Milk"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Dairy"
                                            },
                                            "price": {
                                                "type": "number",
                                                "example": 2.99
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "stock_threshold": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "is_popular": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "is_on_sale": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "sale_price": {
                                                "type": "number",
                                                "example": 1.99,
                                                "nullable": true
                                            },
                                            "rayon": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dairy Section"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}